requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1

google-api-python-client==2.137.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-secret-manager>=2.20.0
SQLAlchemy>=2.0.32

pydantic==2.8.2
python-dotenv==1.0.1

openai
requests
PyPDF2
python-docx
python-docx
PyPDF2
google-cloud-storage>=2.14.0
slack_sdk
python-dotenv
APScheduler==3.10.4
